backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
anthropic>=0.50.0
chromadb>=0.6.0
sentence-transformers>=3.4.0
python-dotenv>=1.0.0
aiofiles>=24.1.0
pypdf>=5.0.0
